requirements.txt
flask 
mysql-connector-python
Flask-Cors